\docType{package}
\name{psgp}
\alias{psgp-package}
\title{Projected Sequential Gaussian Processes for spatial interpolation}
\description{ }
\section{Parameter estimation}{ }
\section{Prediction/Interpolation}{ }
\section{Acknowledgements}{ the PSGP package is written in C++ and uses the IT++ library for all linear algebra routines. }
\references{
  \url{http://sourceforge.net/apps/wordpress/itpp/}
}
